YYC-108: extract tui rendering + terminal init helpers
Problem: src/tui/mod.rs had grown to 2570 lines covering input
routing, palette/picker overlay drawing, slash dispatch, terminal
lifecycle, and the run_tui orchestrator. Reviewers had no way to
read one subsystem without trawling the whole file.

Fix: split the rendering and terminal-lifecycle layers into their
own modules. mod.rs stays as the wiring + orchestrator.

- src/tui/rendering.rs (NEW, 637 lines): all overlay/picker
  drawing helpers — draw_palette, draw_session_picker,
  draw_model_picker, draw_provider_picker, draw_diff_scrubber,
  draw_picker_border (private helper), open_unified_picker,
  initial_path_for_active_model. All marked pub(super) so mod.rs
  calls them via `rendering::draw_*`.
- src/tui/init.rs (NEW, 28 lines): init_terminal +
  restore_terminal — raw mode + alternate screen lifecycle.
- src/tui/mod.rs: 2570 → 1940 lines. Module declarations updated;
  call sites switched to `rendering::` and `init::` qualifiers.
  Unused imports cleaned up (Terminal, Modifier, Style, Line,
  Span, Paragraph, SessionStatus — all migrated with their
  consumers).

Public API unchanged; behavior unchanged. All 355 lib tests pass;
clippy clean.

Out of scope for this PR (follow-ups under YYC-108):
- handle_stream_event + render_wake_for_stream_batch +
  stream_event_forces_redraw → tui/events.rs (~140 lines).
- Slash dispatch (filter_commands, current_palette,
  complete_slash, format_model_list, format_provider_list,
  build_provider_picker_entries, SlashCommand const table) →
  tui/keymap.rs (~250 lines).
- submit_prompt + refresh_sessions glue → keep in mod.rs or
  fold into a new tui/dispatch.rs.

Hitting the issue's <1000-line acceptance for mod.rs needs the
events + keymap extractions; tracked as the YYC-108 follow-up.
